Move @vtex/order-manager to react folder. Update files
