+++ package.json
@@ -26,7 +26,6 @@
   "devDependencies": {
     "@apollo/react-testing": "3",
     "@types/node": "^12.12.21",
-    "@vtex/order-manager": "^0.5.4-beta.0",
     "@vtex/prettier-config": "^0.1.3",
     "@vtex/test-tools": "^3.4.3",
     "eslint": "^6.8.0",
