+++ react/OrderPayment.tsx
@@ -1,6 +1,57 @@
+import { useCallback } from 'react'
+import { OrderQueue } from 'vtex.order-manager'
+import { useMutation } from 'react-apollo'
+import MutationUpdateOrderFormPayment from 'vtex.checkout-resources/MutationUpdateOrderFormPayment'
 import {
-  OrderPaymentProvider,
-  useOrderPayment,
-} from './components/OrderPayment'
+  OrderForm as CheckoutOrderForm,
+  PaymentDataInput,
+} from 'vtex.checkout-graphql'
+import { useOrderForm, useOrderQueue } from '@vtex/order-manager'
 
-export default { useOrderPayment, OrderPaymentProvider }
+import {
+  createOrderPaymentProvider,
+  useOrderPayment,
+} from './components/createOrderPaymentProvider'
+import { useLogger } from './utils/logger'
+
+interface UpdateOrderFormPaymentMutation {
+  updateOrderFormPayment: CheckoutOrderForm
+}
+
+interface UpdateOrderFormPaymentMutationVariables {
+  paymentData: PaymentDataInput
+}
+
+const { useQueueStatus } = OrderQueue
+
+function useUpdateOrderFormPayment() {
+  const [updateOrderFormPayment] = useMutation<
+    UpdateOrderFormPaymentMutation,
+    UpdateOrderFormPaymentMutationVariables
+  >(MutationUpdateOrderFormPayment)
+
+  return {
+    updateOrderFormPayment: useCallback(
+      async (paymentData: PaymentDataInput) => {
+        const { data } = await updateOrderFormPayment({
+          variables: { paymentData },
+        })
+
+        const newOrderForm = data!.updateOrderFormPayment
+
+        return newOrderForm
+      },
+      [updateOrderFormPayment]
+    ),
+  }
+}
+
+const { OrderPaymentProvider } = createOrderPaymentProvider({
+  useLogger,
+  useOrderQueue,
+  useOrderForm,
+  useUpdateOrderFormPayment,
+  useQueueStatus,
+})
+
+export default { OrderPaymentProvider, useOrderPayment }
